lib: improve naming & docs
