+++ lib/propolis/src/lib.rs
@@ -21,9 +21,9 @@ pub mod block;
 pub mod chardev;
 pub mod common;
 pub mod cpuid;
+pub mod enlightenment;
 pub mod exits;
 pub mod firmware;
-pub mod hv_interface;
 pub mod hw;
 pub mod intr_pins;
 pub mod lifecycle;
