+++ lib/propolis/src/vcpu.rs
@@ -11,8 +11,8 @@ use crate::common::GuestData;
 use crate::common::Lifecycle;
 use crate::common::VcpuId;
 use crate::cpuid;
+use crate::enlightenment::Enlightenment;
 use crate::exits::*;
-use crate::hv_interface::Enlightenment;
 use crate::migrate::*;
 use crate::mmio::MmioBus;
 use crate::msr::MsrId;
